MemoryManager should work now and move check_and_eval to misc.py
