+++ qiling/debugger/qdb/context.py
@@ -74,7 +74,7 @@ class Context:
         try to read pointer size of data from ql.mem
         """
 
-        return self.try_read(address, self.pointersize)
+        return self.try_read(address, self.arch_insn_size)
 
     def read_string(self, address: int) -> Optional[str]:
         """
